config
 cho phép đường dẫn từ front-end lấy dữ liệu

+++ src/main/java/com/example/XDMHPL_Back_end/Controller/UserController.java
@@ -30,7 +30,7 @@ public class UserController {
 	}
 	
 	@GetMapping("/{id}")
-	public Users getCustomerById(@PathVariable Long id) {
+	public Users getCustomerById(@PathVariable Integer id) {
 		return usersService.getCustomerById(id);
 	}
